Add aria label identifiers and restructure component
