+++ cardigan/stories/components/Footer/Footer.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, ArgsTable } from '@storybook/addon-docs/blocks';
-import Footer from '@weco/common/views/components/Footer/Footer';
+import Footer from '@weco/common/views/components/Footer';
 import Readme from '@weco/common/views/components/Footer/README.md';
 import * as stories from './Footer.stories.tsx';
