+++ cardigan/stories/components/Footer/Footer.stories.tsx
@@ -1,4 +1,4 @@
-import Footer from '@weco/common/views/components/Footer/Footer';
+import Footer from '@weco/common/views/components/Footer';
 import { venues } from '@weco/common/test/fixtures/components/venues';
 import Readme from '@weco/common/views/components/Footer/README.md';
 import { ReadmeDecorator } from '@weco/cardigan/config/decorators';
