+++ common/views/components/Footer/Footer.WellcomeLogo.tsx
